fix: update endpoints & change tourPreference handling to use userId

+++ src/Modules/Tours/Explorer.Tours.API/Dtos/TourPreferenceDto.cs
@@ -9,8 +9,8 @@ namespace Explorer.Tours.API.Dtos
 {
     public class TourPreferenceDto
     {
-        public int Id { get; set; }
-        public long UserId { get; set; }
+        public long Id { get; set; }
+        public long UserId { get; set; } 
         public TourDifficulty Difficulty { get; set; }
         public TransportType TransportType { get; set; }
         public List<string> Tags { get; set; }
